+++ src/main/java/chess/Application.java
@@ -1,11 +1,11 @@
 package chess;
 
-import chess.controller.ChessController;
+import chess.controller.ConsoleChessController;
 
 public class Application {
 
     public static void main(String[] args) {
-        ChessController chessController = new ChessController();
+        ConsoleChessController chessController = new ConsoleChessController();
         chessController.play();
     }
 }
